backend/requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
sqlparse==0.5.1
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
fuzzywuzzy==0.18.0
sqlparse==0.5.1
